🐛 Fix WhatsApp audio uploads in file inputs
